separate story to make Tab story same as other component
